Server: remove model.py:data_signals threading and use mock
